requirements.txt
streamlit
streamlit_option_menu
python-dotenv
pandas==2.0.3
openpyxl
numpy==1.23.5
plotly==5.15.0
millify
streamlit_extras==0.3.0
altair==5.0.1
xlrd
google-generativeai
scikit-learn
chardet
